App builder is now functional
